small refactor for actuator folder
